Rename SVD models
