+++ lenskit-mf/src/main/java/org/lenskit/mf/MFModel.java
@@ -22,7 +22,7 @@
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.lenskit.mf.svd;
+package org.lenskit.mf;
 
 import com.google.common.base.Preconditions;
 import org.apache.commons.math3.linear.MatrixUtils;
